dspback-django/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
djangorestframework-api-key>=3.0,<4.0
django-cors-headers>=4.3,<5.0
django-allauth[socialaccount]>=65.0,<66.0
psycopg2-binary>=2.9,<3.0
jsonschema[format]>=4.23,<5.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
requests>=2.31,<3.0
